rename open_phone client to service
