+++ api/src/open_phone/routes.py
@@ -12,7 +12,7 @@ from pydantic import BaseModel
 from api.src.database.database import get_session
 from api.src.open_phone.models import OpenPhoneEvent
 from api.src.open_phone.schema import OpenPhoneWebhookPayload
-from api.src.open_phone.client import send_message, get_contacts_by_external_ids, get_contacts_sheet_as_json
+from api.src.open_phone.service import send_message, get_contacts_by_external_ids, get_contacts_sheet_as_json
 from api.src.open_phone.escalate import analyze_for_twilio_escalation
 from api.src.utils.password import verify_admin_auth
 import asyncio
